+++ extensions/blog/src/Controller/PostController.php
@@ -153,6 +153,8 @@ class PostController extends Controller
             }
 
             $data['date'] = $this['dates']->getDateTime($data['date'])->setTimezone(new \DateTimeZone('UTC'));
+            
+            $data['comment_status'] = isset($data['comment_status']) ? $data['comment_status'] : 0;
 
             $this->posts->save($post, $data);